chore(edit): remove entur theme
