+++ next-tavla/app/(admin)/edit/[id]/components/MetaSettings/actions.ts
@@ -10,7 +10,7 @@ import admin, { firestore } from 'firebase-admin'
 import { revalidatePath } from 'next/cache'
 import { redirect } from 'next/navigation'
 import { TFontSize, TLocation } from 'types/meta'
-import { TBoard, TBoardID, TOrganizationID, TTheme } from 'types/settings'
+import { TBoard, TBoardID, TOrganizationID } from 'types/settings'
 import { getBoard, getWalkingDistanceTile } from '../../actions'
 import { getUserFromSessionCookie } from 'app/(admin)/utils/server'
 
@@ -105,18 +105,6 @@ export async function moveBoard(
             .collection('users')
             .doc(user.uid)
             .update({ owner: admin.firestore.FieldValue.arrayUnion(bid) })
-
-    revalidatePath(`/edit/${bid}`)
-}
-
-export async function setTheme(bid: TBoardID, theme?: TTheme) {
-    const access = await hasBoardEditorAccess(bid)
-    if (!access) return redirect('/')
-
-    await firestore()
-        .collection('boards')
-        .doc(bid)
-        .update({ theme: theme ?? 'dark', 'meta.dateModified': Date.now() })
 
     revalidatePath(`/edit/${bid}`)
 }
