+++ next-tavla/app/(admin)/edit/[id]/page.tsx
@@ -17,7 +17,7 @@ import { RefreshButton } from './components/RefreshButton'
 import { DEFAULT_BOARD_NAME } from 'app/(admin)/utils/constants'
 import { Preview } from './components/Preview'
 import { ActionsMenu } from './components/ActionsMenu'
-import { ThemeSelect } from './components/ThemeSwitch'
+import { ThemeSwitch } from './components/ThemeSwitch'
 
 type TProps = {
     params: { id: TBoardID }
@@ -63,7 +63,7 @@ export default async function EditPage({ params }: TProps) {
                     footer={board.footer}
                     organizationBoard={organization !== undefined}
                 />
-                <ThemeSelect board={board} />
+                <ThemeSwitch board={board} />
             </div>
 
             <div className="flex flex-col gap-4">
